+++ CHANGELOG.md
@@ -12,6 +12,19 @@ The Call Message Events (Beta), originally released in 2.2.0, has been promoted 
 - A new error, [31210](https://www.twilio.com/docs/api/errors/31210), has been added to the SDK. This new error is emitted via [call.on('error', handler(twilioError))](https://twilio.github.io/twilio-voice.js/classes/voice.call.html#errorevent) after calling the [sendMessage](https://twilio.github.io/twilio-voice.js/classes/voice.call.html#sendmessage) API with an invalid [Call.Message.messageType](https://twilio.github.io/twilio-voice.js/interfaces/voice.call.message.html).
 - Fixed an issue where the wrong error code `31209` is raised if the payload size of a Call Message Event exceeds the authorized limit. With this release, `31212` is raised instead.
 
+2.11.3 (August 21, 2024)
+========================
+
+Bug Fixes
+---------
+
+- Fixed an issue where `PreflightTest` throws an error when `RTCIceCandidateStatsReport` is not available. Thanks @phi-line for your [contribution](https://github.com/twilio/twilio-voice.js/pull/280).
+
+Improvements
+------------
+
+- The SDK now updates its internal device list when the microphone permission changes.
+
 2.11.2 (June 26, 2024)
 ======================